Stop propagating clicks from tooltip popper
Previously the click event would propagate to the parent element of the
tooltip, which would be unexpected and undesirable in most cases.
